fix: Resolve conflicts
